% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_opt.R
\name{EstOptions}
\alias{EstOptions}
\title{Default Estimation Options}
\usage{
EstOptions()
}
\value{
A list object \code{opt} that stores the default estimation options.
}
\description{
\code{EstOptions} defines the estimation options used in the FCVAR
  estimation procedure and the related programs.
}
\examples{
EstOptions()
}
\seealso{
\code{updateRestrictions} to set and test estimation options for validity and compatibility.
\code{FCVARestn} for use of these options in estimation.

Other FCVAR option functions: 
\code{\link{GetBounds}()},
\code{\link{updateRestrictions}()}
}
\concept{FCVAR option functions}
