% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR.R
\docType{package}
\name{FCVAR}
\alias{FCVAR}
\title{FCVAR: A package for estimating the Fractionally Cointegrated VAR model.}
\description{
The FCVAR package provides three categories of important functions:
higher, lower and estn.
}
\section{FCVAR functions}{

The FCVAR functions ...
}

