% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{FCVARforecast}
\alias{FCVARforecast}
\title{Forecasts with the FCVAR Model}
\usage{
FCVARforecast(x, model, NumPeriods)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.
The forecast will be calculated using these values as starting values.}

\item{model}{A list of estimation results, just as if estimated from \code{FCVARest}.
The parameters in \code{model} can also be set or adjusted by assigning new values.}

\item{NumPeriods}{The number of time periods in the simulation.}
}
\value{
A \code{NumPeriods} \eqn{\times p} matrix \code{xf} of forecasted values.
}
\description{
\code{FCVARforecast} calculates recursive forecasts with the FCVAR model.
}
\examples{
opt <- EstOptions()
x <- data(JNP2014)
model <- FCVARestn(x,k = 3,r = 1,opt)
xf <- FCVARforecast(data, model, NumPeriods = 100)
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} for the specification of the \code{model}.
\code{FCVARforecast} calls \code{FracDiff} and \code{Lbk} to calculate the forecast.

Other FCVAR auxilliary functions: 
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{FreeParams}()},
\code{\link{FullFCVARlike}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{RstrctOptm_Switch}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
