% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{FCVARlike}
\alias{FCVARlike}
\title{Likelihood Function for the Constrained FCVAR Model}
\usage{
FCVARlike(params, x, k, r, opt)
}
\arguments{
\item{params}{A vector of parameters \code{d} and \code{b}
(and \code{mu} if option selected).}

\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A number \code{like}, the log-likelihood evaluated at the
specified parameter values.
}
\description{
\code{FCVARlike} calculates the likelihood for the constrained FCVAR model
for a given set of parameter values. This function adjusts the variables with the level parameter,
if present, and returns the log-likelihood given \code{d} and \code{b}.
}
\examples{
opt <- FCVARoptions()
x <- data(JNP2014)
like <- FCVARlike(x, params = c(1, 1), , k = 2, r = 1, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{FreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{RstrctOptm_Switch}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
