% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{FCVARlikeFull}
\alias{FCVARlikeFull}
\title{Likelihood Function for the FCVAR Model}
\usage{
FCVARlikeFull(x, k, r, coeffs, betaHat, rhoHat, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{coeffs}{A list of coefficients of the FCVAR model.
It is an element of the list \code{results} returned by \code{FCVARestn},
without the parameters \code{betaHat} and \code{rhoHat}.}

\item{betaHat}{A \eqn{p x r} matrix of cointegrating vectors.
The \eqn{r x 1} vector \eqn{\beta x_t} is the stationary cointegration relations.}

\item{rhoHat}{A \eqn{p x 1} vector of restricted constatnts.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A number \code{like}, the log-likelihood evaluated at the
specified parameter values.
}
\description{
\code{FCVARlikeFull} calculates the likelihood for the constrained FCVAR model
for a given set of parameter values.
This function takes the full set of coefficients from a call
to \code{FCVARestn} and returns the log-likelihood given \code{d} and \code{b}.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
FCVARlikeFull(x, k = 2, r = 1, coeffs = results$coeffs,
              beta = results$coeffs$betaHat, rho = results$coeffs$rhoHat, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} for the estimation of coefficients in \code{coeffs}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeGrid}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetEstimates}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{Lbk}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()},
\code{\link{plot.FCVAR_grid}()}
}
\concept{FCVAR auxilliary functions}
