% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{FCVARlikeMu}
\alias{FCVARlikeMu}
\title{Likelihood Function for the Unconstrained FCVAR Model}
\usage{
FCVARlikeMu(mu, y, db, k, r, opt)
}
\arguments{
\item{mu}{The level parameter.}

\item{y}{A matrix of variables to be included in the system.}

\item{db}{The orders of fractional integration.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A number \code{like}, the log-likelihood evaluated at the
specified parameter values.
}
\description{
\code{FCVARlikeMu} calculates the likelihood for the unconstrained FCVAR model
for a given set of parameter values. It is used by the \code{FCVARlikeGrid}
function to numerically optimize over the level parameter for given values of
the fractional parameters.
}
\examples{
opt <- FCVARoptions()
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
like <- FCVARlikeMu(mu = colMeans(x), y = x, db = c(1, 1), k = 2, r = 1, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
The \code{FCVARlikeGrid} calls this function to perform a grid search over the
parameter values.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeGrid}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetEstimates}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{Lbk}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()},
\code{\link{plot.FCVAR_grid}()}
}
\concept{FCVAR auxilliary functions}
