% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_opt.R
\name{FCVARoptions}
\alias{FCVARoptions}
\title{Default Estimation Options}
\usage{
FCVARoptions()
}
\value{
An S3 object of class \code{FCVAR_opt} that stores the default estimation options.
}
\description{
\code{FCVARoptions} defines the estimation options used in the FCVAR
  estimation procedure and the related programs.
}
\examples{
opt <- FCVARoptions()
}
\seealso{
\code{FCVARoptionUpdates} to set and test estimation options for validity and compatibility.
\code{FCVARestn} for use of these options in estimation.

Other FCVAR estimation functions: 
\code{\link{FCVARestn}()},
\code{\link{FCVARoptionUpdates}()},
\code{\link{GetBounds}()},
\code{\link{summary.FCVAR_model}()}
}
\concept{FCVAR estimation functions}
