% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_opt.R
\name{FCVARoptionUpdates}
\alias{FCVARoptionUpdates}
\title{Update Estimation Options for FCVAR Restrictions}
\usage{
FCVARoptionUpdates(opt, p, r)
}
\arguments{
\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}

\item{p}{The number of variables in the system.}

\item{r}{The cointegrating rank.}
}
\value{
A list object \code{newOpt} that stores the chosen estimation options,
with appropriate updates based on user-defined options.
}
\description{
A function to set and test estimation options for validity and compatibility.
This function is called prior to estimation and
	performs several checks to verify the input in the estimation
	options:
\itemize{
  \item Check that only one option for deterministics is chosen,
  \item Check that starting values have been specified correctly,
  \item Update \code{dbMin}, \code{dbMax}, based on user-specified options,
  \item Check for appropriate dimensions and redundancies in the restriction matrices \code{R_psi}, \code{R_Alpha} and \code{R_Beta}.
}
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
newOpt <- FCVARoptionUpdates(opt, p = 3, r = 1)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} calls this function at the start of each estimation to verify
validity of options.

Other FCVAR estimation functions: 
\code{\link{FCVARestn}()},
\code{\link{FCVARoptions}()},
\code{\link{GetBounds}()},
\code{\link{summary.FCVAR_model}()}
}
\concept{FCVAR estimation functions}
