% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{FCVARsimBS}
\alias{FCVARsimBS}
\title{Draw Bootstrap Samples from the FCVAR Model}
\usage{
FCVARsimBS(data, model, NumPeriods)
}
\arguments{
\item{data}{A \eqn{T \times p} matrix of starting values for the simulated realizations.}

\item{model}{A list of estimation results, just as if estimated from \code{FCVARest}.
The parameters in \code{model} can also be set or adjusted by assigning new values.}

\item{NumPeriods}{The number of time periods in the simulation.}
}
\value{
A \code{NumPeriods} \eqn{\times p} matrix \code{xBS} of simulated bootstrap values.
}
\description{
\code{FCVARsimBS} simulates the FCVAR model as specified by
input \code{model} and starting values specified by \code{data}.
It creates a bootstrap sample by augmenting each iteration
with a bootstrap error. The errors are sampled from the
residuals specified under the \code{model} input and have a
positive or negative sign with equal probability (the Rademacher distribution).
}
\examples{
opt <- EstOptions()
x <- data(JNP2014)
model <- FCVARestn(x,k = 3,r = 1,opt)
data <- x[1:10, ]
xBS <- FCVARsimBS(data, model, NumPeriods = 100)
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} for the specification of the \code{model}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FullFCVARlike}()},
\code{\link{GetParams}()},
\code{\link{LikeGrid}()}
}
\concept{FCVAR auxilliary functions}
