% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{find_local_max}
\alias{find_local_max}
\title{Find local optima}
\usage{
find_local_max(x)
}
\arguments{
\item{x}{A 1- or 2-dimensional numerical array of the candidate values
of the objective function.}
}
\value{
A list object with three elements:
\describe{
  \item{\code{row}}{An integer row number of the local maxima. }
  \item{\code{col}}{An integer column number of the local maxima. }
  \item{\code{value}}{The numeric values of the local maxima. }
}
}
\description{
\code{find_local_max} finds the row and column numbers corresponding
to the local maxima of a 1- or 2-dimensional array.
}
