% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{FracDiff}
\alias{FracDiff}
\title{Fast Fractional Differencing}
\usage{
FracDiff(x, d)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{d}{The order of fractional differencing.}
}
\value{
A vector or matrix \code{dx} equal to \eqn{(1-L)^d x}
of the same dimensions as x.
}
\description{
\code{FracDiff} is a fractional differencing procedure based on the
	fast fractional difference algorithm of Jensen & Nielsen (2014).
}
\examples{
opt <- FCVARoptions()
x <- data(JNP2014)
dx = FracDiff(x, d = 0.5)
}
\references{
Jensen, A. N. and M. \enc{Ø}{O}. Nielsen (2014).
"A fast fractional difference algorithm,"
Journal of Time Series Analysis 35, 428-436.
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} calls \code{GetParams}, which calls \code{TransformData}
to estimate the FCVAR model.
\code{TransformData} in turn calls \code{FracDiff} and \code{Lbk}
to perform the transformation.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{RstrctOptm_Switch}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
