% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{FullFCVARlike}
\alias{FullFCVARlike}
\title{Likelihood Function for the FCVAR Model}
\usage{
FullFCVARlike(x, k, r, coeffs, beta, rho, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{coeffs}{A list of coefficients of the FCVAR model.
It is an element of the list \code{results} returned by \code{FCVARestn},
without the parameters \code{betaHat} and \code{rhoHat}.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{EstOptions()}.}

\item{betaHat}{A \eqn{p x r} matrix of cointegrating vectors.
The \eqn{r x 1} vector \eqn{\betax_t} is the stationary cointegration relations.}

\item{rhoHat}{A \eqn{p x 1} vector of restricted constatnts.}
}
\value{
A number \code{like}, the log-likelihood evaluated at the
specified parameter values.
}
\description{
\code{FullFCVARlike} calculates the likelihood for the constrained FCVAR model
for a given set of parameter values.
This function takes the full set of coefficients from a call
to \code{FCVARestn} and returns the log-likelihood given \code{d} and \code{b}.
}
\examples{
opt <- EstOptions()
x <- data(JNP2014)
results <- FCVARestn(x, k = 2, r = 1, opt)
like <- FullFCVARlike(x, k = 2, r = 1, coeffs = results$coeffs,
beta = coeffs$betaHat, rho = coeffs$rhoHat, opt)
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} for the estimation of coefficients in \code{coeffs}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
