% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_opt.R
\name{GetBounds}
\alias{GetBounds}
\title{Obtain Bounds on FCVAR Parameters}
\usage{
GetBounds(opt)
}
\arguments{
\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A list object \code{UB_LB_bounds} containing the upper (\code{UB}) and lower (\code{LB})  bounds on either \code{db} or \code{phi}.
These vectors are either 1- or 2-dimensional depending on whether a restriction on \code{d} and \code{b} is imposed or not.
}
\description{
\code{GetBounds()} obtains bounds on fractional integration parameters in the FCVAR model.
  This function obtains upper and lower bounds on \code{d}, \code{b} or on
  \code{phi}, which is given by \code{db <- H*phi + h},
  using the restrictions implied by \code{H} and \code{h}.
}
\examples{
opt <- FCVARoptions()
UB_LB_bounds <- GetBounds(opt)

opt <- FCVARoptions()
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
UB_LB_bounds <- GetBounds(opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} calls this function at the start of each estimation to specify any bounds on fractional integration parameters.

Other FCVAR estimation functions: 
\code{\link{FCVARestn}()},
\code{\link{FCVARoptionUpdates}()},
\code{\link{FCVARoptions}()},
\code{\link{summary.FCVAR_model}()}
}
\concept{FCVAR estimation functions}
