% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{GetCharPolyRoots}
\alias{GetCharPolyRoots}
\title{Roots of the Characteristic Polynomial}
\usage{
GetCharPolyRoots(coeffs, opt, k, r, p)
}
\arguments{
\item{coeffs}{A list of coefficients for the FCVAR model.
An element of the list of estimation \code{results} output from \code{FCVARestn}.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{EstOptions()}.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{p}{The number of variables in the system.}
}
\value{
A complex vector \code{cPolyRoots} with the roots of the characteristic polynomial.
}
\description{
\code{GetCharPolyRoots} calculates the roots of the
characteristic polynomial and plots them with the unit circle
transformed for the fractional model, see Johansen (2008).
}
\note{
The roots are calculated from the companion form of the VAR,
where the roots are given as the inverse eigenvalues of the
coefficient matrix.
}
\examples{
opt <- EstOptions()
optRES$R_Beta <- matrix(c(1, 0, 0), nrow = 1, ncol = 3)
x <- data(JNP2014)
results <- FCVARestn(x, k = 2,r = 1, opt)
cPolyRoots <- GetCharPolyRoots(results$coeffs, opt, k = 2, r = 1, p = 3)
}
\references{
Johansen, S. (2008). "A representation theory for a class of
vector autoregressive models for fractional processes,"
Econometric Theory 24, 651-676.
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} to estimate the model for which to calculate the roots
of the characteristic polynomial.

Other FCVAR postestimation functions: 
\code{\link{FCVARboot}()},
\code{\link{HypoTest}()},
\code{\link{LMtest}()},
\code{\link{Qtest}()},
\code{\link{mv_wntest}()}
}
\concept{FCVAR postestimation functions}
