% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{GetEstimates}
\alias{GetEstimates}
\title{Obtain concentrated parameter estimates for the Constrained FCVAR Model}
\usage{
GetEstimates(params, x, k, r, opt)
}
\arguments{
\item{params}{A vector of parameters \code{d} and \code{b}
(and \code{mu} if option selected).}

\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A list object \code{estimates} containing the estimates,
including the following parameters:
\describe{
  \item{\code{db}}{The orders of fractional integration (taken directly from the input).}
  \item{\code{alphaHat}}{A \eqn{p x r} matrix of adjustment parameters.}
  \item{\code{betaHat}}{A \eqn{p x r} matrix of cointegrating vectors.
      The \eqn{r x 1} vector \eqn{\beta x_t} is the stationary cointegration relations.}
  \item{\code{rhoHat}}{A \eqn{p x 1} vector of restricted constatnts.}
  \item{\code{piHat}}{A \eqn{p x p} matrix \eqn{\Pi = \alpha \beta'} of long-run levels.}
  \item{\code{OmegaHat}}{A \eqn{p x p} covariance matrix of the error terms.}
  \item{\code{GammaHat}}{A ( \eqn{p x kp} matrix \code{cbind(GammaHat1,...,GammaHatk)})
  of autoregressive coefficients. }
  \item{\code{muHat}}{A vector of the optimal \code{mu} if level parameter is selected. }
}
}
\description{
\code{GetEstimates} calculates the concentrated parameter estimates for
the constrained FCVAR model for a given set of parameter values.
It is used after estimating the parameters that are numerically optimized
to obtain the corresponding concentrated parameters.
Like \code{FCVARlike} , this function adjusts the variables with the level parameter,
if present, and returns the log-likelihood given \code{d} and \code{b}.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
GetEstimates(c(results$coeffs$db, results$coeffs$muHat), x, k = 2, r = 1, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARlike} performs the same calculations to obtain the value
of the likelihood function.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeGrid}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{Lbk}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()},
\code{\link{plot.FCVARlikeGrid}()}
}
\concept{FCVAR auxilliary functions}
