% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{GetParams}
\alias{GetParams}
\title{Calculate Parameters of the Cointegrated VAR Model}
\usage{
GetParams(x, k, r, db, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{db}{The orders of fractional integration.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{EstOptions()}.}
}
\value{
A list object \code{estimates} containing the estimates,
including the following parameters:
\describe{
  \item{\code{db}}{The orders of fractional integration (taken directly from the input).}
  \item{\code{alphaHat}}{A \eqn{p x r} matrix of adjustment parameters.}
  \item{\code{betaHat}}{A \eqn{p x r} matrix of cointegrating vectors.
      The \eqn{r x 1} vector \eqn{\beta x_t} is the stationary cointegration relations.}
  \item{\code{rhoHat}}{A \eqn{p x 1} vector of restricted constatnts.}
  \item{\code{piHat}}{A \eqn{p x p} matrix \eqn{\Pi = \alpha \beta'} of long-run levels.}
  \item{\code{OmegaHat}}{A \eqn{p x p} covariance matrix of the error terms.}
  \item{\code{GammaHat}}{A ( \eqn{p x kp} matrix \code{cbind(GammaHat1,...,GammaHatk)})
  of autoregressive coefficients. }
}
}
\description{
\code{GetParams} calculates the parameters of the cointegrated VAR model
 from a multivariate sample.
 This function uses FWL and reduced rank regression to obtain
the estimates of \code{Alpha}, \code{Beta}, \code{Rho}, \code{Pi},
\code{Gamma}, and \code{Omega}, using the notation in Johansen and Nielsen (2012).
}
\examples{
opt <- EstOptions()
x <- data(JNP2014)
estimates <- GetParams(x, k = 2, r = 1, db = c(1, 1), opt)
}
\references{
Johansen, S. and M. \enc{Ø}{O}. Nielsen (2012).
"Likelihood inference for a fractionally cointegrated
vector autoregressive model," Econometrica 80, 2667-2732.
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} calls \code{GetParams} to estimate the FCVAR model.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{FullFCVARlike}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
