% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_spec.R
\name{GetPvalues}
\alias{GetPvalues}
\title{Get p-values for Unit Roots and Cointegration}
\usage{
GetPvalues(q, b, consT, testStat, opt)
}
\arguments{
\item{q}{Number of variables, minus the cointegrating rank.}

\item{b}{Fractional integration parameter.}

\item{testStat}{Value of the test statistic.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}

\item{const}{Boolean variable indicating whether or not there is a constant present.}
}
\value{
A scalar numeric \code{pv}, the p-value for the likelihood ratio test.
}
\description{
Calculates p-values for tests of fractional unit roots and cointegration
This function calls the program FDPVAL in the terminal and
	returns the p-value based on the user's inputs. The function's
	arguments must be converted to strings in order to interact with the
	terminal.
}
\section{Non-stationary Systems}{

For non-stationary systems \code{b >= 0.5}, use simulated P-values from
MacKinnon and Nielsen (2014, Journal of Applied Econometrics)
and the C++ program conversion by Jason Rhinelander.
Note: fdpval is a separately installed program.
For more information see: \url{https://github.com/jagerman/fracdist}
For download see \url{https://github.com/jagerman/fracdist/releases}
}

\section{Stationary Systems}{

For stationary systems \code{b < 0.5}, use chi^2 with \code{(p-r)^2}
degrees of freedom. See Johansen and Nielsen (2012, Econometrica).
}

\examples{
opt <- FCVARoptions()
pv <- GetPvalues(q = 1, b = 0.4, consT = 0, testStat = 3.84, opt)

opt <- FCVARoptions()
pv <- GetPvalues(q = 1, b = 0.75, consT = 0, testStat = 3.84, opt)
}
\references{
James G. MacKinnon and Morten \enc{Ø}{O}rregaard Nielsen,
"Numerical Distribution Functions of Fractional Unit Root and Cointegration Tests,"
Journal of Applied Econometrics, Vol. 29, No. 1, 2014, pp.161-171.

Johansen, S. and M. \enc{Ø}{O}. Nielsen (2012).
"Likelihood inference for a fractionally cointegrated vector autore-gressive model,"
Econometrica 80, pp.2667-2732.
}
\seealso{
\code{FCVARoptions} to set default estimation options.

Other FCVAR specification functions: 
\code{\link{FCVARbootRank}()},
\code{\link{LagSelect}()},
\code{\link{RankTests}()},
\code{\link{print.LagSelect}()}
}
\concept{FCVAR specification functions}
