% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{GetResiduals}
\alias{GetResiduals}
\title{Calculate Residuals for the FCVAR Model}
\usage{
GetResiduals(x, k, r, coeffs, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{coeffs}{A list of coefficients of the FCVAR model.
It is an element of the list \code{results} returned by \code{FCVARestn}.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A matrix \code{epsilon} of residuals from FCVAR model estimation
calculated with the parameter estimates specified in \code{coeffs}.
}
\description{
\code{GetResiduals} calculates residuals for the FCVAR model
from given parameter values.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
epsilon <- GetResiduals(x, k = 2, r = 1, coeffs = results$coeffs, opt)
}
\references{
Johansen, S. and M. \enc{Ø}{O}. Nielsen (2012).
"Likelihood inference for a fractionally cointegrated
vector autoregressive model," Econometrica 80, 2667-2732.
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} to estimate the FCVAR model.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetEstimates}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetRestrictedParams}()},
\code{\link{Lbk}()},
\code{\link{LikeGridSearch}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()},
\code{\link{plot.LikeGridSearch}()}
}
\concept{FCVAR auxilliary functions}
