% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{GetRestrictedParams}
\alias{GetRestrictedParams}
\title{Calculate Restricted Estimates for the FCVAR Model}
\usage{
GetRestrictedParams(beta0, S00, S01, S11, cap_T, p, opt)
}
\arguments{
\item{beta0}{The unrestricted estimate of \code{beta},
a \eqn{p x r} matrix of cointegrating vectors
returned from \code{FCVARestn} or \code{GetParams}.}

\item{S00}{A matrix of product moments,
calculated from the output of \code{TransformData} in \code{GetParams}.}

\item{S01}{A matrix of product moments,
calculated from the output of \code{TransformData} in \code{GetParams}.}

\item{S11}{A matrix of product moments,
calculated from the output of \code{TransformData} in \code{GetParams}.}

\item{cap_T}{The number of observations in the sample.}

\item{p}{The number of variables in the system.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.

#' @return A list object \code{switched_mats} containing the restricted estimates,
including the following parameters:
\describe{
  \item{\code{betaStar}}{A \eqn{p x r} matrix of cointegrating vectors.
      The \eqn{r x 1} vector \eqn{\beta x_t} is the stationary cointegration relations.}
  \item{\code{alphaHat}}{A \eqn{p x r} matrix of adjustment parameters.}
  \item{\code{OmegaHat}}{A \eqn{p x p} covariance matrix of the error terms.}
}}
}
\description{
\code{GetRestrictedParams} calculates restricted estimates of
cointegration parameters and error variance in the FCVAR model.
To calculate the optimum, it uses the switching algorithm
of Boswijk and Doornik (2004, page 455) to optimize over free parameters
\eqn{\psi} and \eqn{\phi} directly, combined with the line search proposed by
Doornik (2016, working paper). We translate between  \eqn{(\psi, \phi)} and
\eqn{(\alpha, \beta)} using the relation of \eqn{R_\alpha vec(\alpha) = 0} and
\eqn{A\psi = vec(\alpha')}, and \eqn{R_\beta vec(\beta) = r_\beta} and
\eqn{H\phi + h = vec(\beta)}. Note the transposes.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
opt$R_Beta <- matrix(c(1, 0, 0), nrow = 1, ncol = 3)
opt <- FCVARoptionUpdates(opt, p = 3, r = 1) # Need to update restriction matrices.
betaStar <- matrix(c(-0.95335616, -0.07345676, -0.29277318), nrow = 3)
S00 <- matrix(c(0.0302086527,  0.001308664,  0.0008200923,
                0.0013086640,  0.821417610, -0.1104617893,
                0.0008200923, -0.110461789,  0.0272861128), nrow = 3)
S01 <- matrix(c(-0.0047314320, -0.04488533,  0.006336798,
                0.0026708007,  0.17463884, -0.069006455,
                -0.0003414163, -0.07110324,  0.022830494), nrow = 3, byrow = TRUE)
S11 <- matrix(c( 0.061355941,  -0.4109969,  -0.007468716,
                 -0.410996895,  70.6110313, -15.865097810,
                 -0.007468716, -15.8650978,   3.992435799), nrow = 3)
switched_mats <- GetRestrictedParams(betaStar, S00, S01, S11, cap_T = 316, p = 3, opt)
}
\references{
Boswijk, H. P. and J. A. Doornik (2004).
"Identifying, estimating and testing restricted cointegrated systems:
An overview," Statistica Neerlandica 58, 440-465.

Doornik, J. A. (2018).
"Accelerated estimation of switching algorithms: the cointegrated
VAR model and other applications,"
Forthcoming in Scandinavian Journal of Statistics.
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} calls \code{GetParams} to estimate the FCVAR model,
which in turn calls \code{GetRestrictedParams} if there are restrictions
imposed on \code{alpha} or \code{beta}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeGrid}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetEstimates}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()},
\code{\link{plot.FCVAR_grid}()}
}
\concept{FCVAR auxilliary functions}
