% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{HypoTest}
\alias{HypoTest}
\title{Test of Restrictions on FCVAR Model}
\usage{
HypoTest(modelUNR, modelR)
}
\arguments{
\item{modelUNR}{A list of estimation results created for the unrestricted model.}

\item{modelR}{A list of estimation results created for the restricted model.}
}
\value{
A list object \code{LRtest} containing the test results,
including the following parameters:
\describe{
  \item{\code{loglikUNR}}{The log-likelihood for the unrestricted model.}
  \item{\code{loglikR}}{The log-likelihood for the restricted model.}
  \item{\code{df}}{The degrees of freedom for the test.}
  \item{\code{LRstat}}{The likelihood ratio test statistic.}
  \item{\code{p_LRtest}}{The p-value for the likelihood ratio test.}
}
}
\description{
\code{HypoTest} performs a likelihood ratio test of the null
	hypothesis: "model is \code{modelR}" against the alternative hypothesis:
	"model is \code{modelUNR}".
}
\examples{
opt <- FCVARoptions()
x <- data(JNP2014)
modelUNR <- FCVARestn(x,k = 2,r = 1, opt)
opt1 <- opt
# Test a hypothesis on the cointegrating vector.
opt1$R_Beta <- matrix(c(1, 0, 0), nrow = 1, ncol = 3)
modelR <- FCVARestn(x1, k, r, opt1)
Hbeta <- HypoTest(modelUNR, modelR)
}
\seealso{
The test is calculated using the results of two calls to
\code{FCVARestn}, under the restricted and unrestricted models.
Use \code{FCVARoptions} to set default estimation options for each model,
then set restrictions as needed before \code{FCVARestn}.

Other FCVAR postestimation functions: 
\code{\link{FCVARboot}()},
\code{\link{GetCharPolyRoots}()},
\code{\link{LMtest}()},
\code{\link{Qtest}()},
\code{\link{mv_wntest}()}
}
\concept{FCVAR postestimation functions}
