% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{Lbk}
\alias{Lbk}
\title{Calculate Lag Polynomial in the Fractional Lag Operator}
\usage{
Lbk(x, b, k)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{b}{The order of fractional differencing.}

\item{k}{The number of lags in the system.}
}
\value{
A matrix \code{Lbkx} of the form \eqn{[ Lb^1 x, Lb^2 x, ..., Lb^k x]}
where \eqn{Lb = 1 - (1-L)^b}.
The output matrix has the same number of rows as \code{x}
but \code{k} times as many columns.
}
\description{
\code{Lbk} calculates a lag polynomial in the fractional lag operator.
}
\examples{
opt <- FCVARoptions()
x <- data(JNP2014)
Lbkx <- Lbk(x, b = 0.5, k = 2)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} calls \code{GetParams}, which calls \code{TransformData}
to estimate the FCVAR model.
\code{TransformData} in turn calls \code{FracDiff} and \code{Lbk}
to perform the transformation.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{LikeGridSearch}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
