% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{LikeGrid}
\alias{LikeGrid}
\title{Grid Search over Likelihood Values}
\usage{
LikeGrid(x, k, r, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A vector \code{params} of \code{d} and \code{b}
(and \code{mu} if level parameter is selected)
corresponding to a maximum over the grid of \code{c(d,b)} or \code{phi}.
}
\description{
\code{LikeGrid} calculates the likelihood function
on a grid of candidate parameter values.
This function evaluates the likelihood over a grid of values
	for \code{c(d,b)} (or \code{phi}).
	It can be used when parameter estimates are sensitive to
	starting values to give an approximation of the global max which can
	then be used as the starting value in the numerical optimization in
	\code{FCVARestn}.
}
\note{
If \code{opt$LocalMax == 0}, \code{LikeGrid} returns the parameter values
      corresponding to the global maximum of the likelihood on the grid.
      If \code{opt$LocalMax == 1}, \code{LikeGrid} returns the parameter values for the
      local maximum corresponding to the highest value of \code{b}. This
      alleviates the identification problem mentioned in Johansen and
      Nielsen (2010, section 2.3).
}
\examples{
opt <- FCVARoptions()
x <- data(JNP2014)
params <- LikeGrid(x, k = 2, r = 1, opt)
}
\references{
Johansen, S. and M. \enc{Ø}{O}. Nielsen (2010).
"Likelihood inference for a nonstationary fractional
autoregressive model," Journal of Econometrics 158, 51-66.
}
\seealso{
\code{FCVARoptions} to set default estimation options.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{FreeParams}()},
\code{\link{FullFCVARlike}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{RstrctOptm_Switch}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
