% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{LikeGridSearch}
\alias{LikeGridSearch}
\title{Grid Search to Maximize Likelihood Function}
\usage{
LikeGridSearch(x, k, r, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A list object \code{likeGrid_params} containing the optimization results,
including the following parameters:
\describe{
  \item{\code{params}}{A vector \code{params} of \code{d} and \code{b}
(and \code{mu} if level parameter is selected)
corresponding to a maximum over the grid of \code{c(d,b)} or \code{phi}.}
  \item{\code{dbHatStar}}{A vector of \code{d} and \code{b}
  corresponding to a maximum over the grid of \code{c(d,b)} or \code{phi}}
  \item{\code{muHatStar}}{A vector of the optimal \code{mu} if level parameter is selected. }
  \item{\code{Grid2d}}{An indicator for whether or not the optimization
  is conducted over a 2-dimensional parameter space,
  i.e. if there is no equality restriction on \code{d} and \code{b}.}
  \item{\code{dGrid}}{A vector of the grid points in the parameter \code{d}.}
  \item{\code{bGrid}}{A vector of the grid points in the parameter \code{b}.}
  \item{\code{like}}{The maximum value of the likelihood function over the chosen grid.}
}
}
\description{
\code{LikeGridSearch} performs a grid-search optimization
by calculating the likelihood function
on a grid of candidate parameter values.
This function evaluates the likelihood over a grid of values
	for \code{c(d,b)} (or \code{phi}).
	It can be used when parameter estimates are sensitive to
	starting values to give an approximation of the global max which can
	then be used as the starting value in the numerical optimization in
	\code{FCVARestn}.
	\code{print.LikeGridSearch} plots the likelihood function from \code{LikeGridSearch}.
}
\note{
If \code{opt$LocalMax == 0}, \code{LikeGridSearch} returns the parameter values
      corresponding to the global maximum of the likelihood on the grid.
      If \code{opt$LocalMax == 1}, \code{LikeGridSearch} returns the parameter values for the
      local maximum corresponding to the highest value of \code{b}. This
      alleviates the identification problem mentioned in Johansen and
      Nielsen (2010, section 2.3).
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
opt$progress <- 2 # Show progress report on each value of b.
newOpt <- FCVARoptionUpdates(opt, p = 3, r = 1) # Need to update restriction matrices.
likeGrid_params <- LikeGridSearch(x, k = 2, r = 1, newOpt)
}
\references{
Johansen, S. and M. \enc{Ø}{O}. Nielsen (2010).
"Likelihood inference for a nonstationary fractional
autoregressive model," Journal of Econometrics 158, 51-66.
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{print.LikeGridSearch} plots the likelihood function from \code{LikeGridSearch}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{Lbk}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()},
\code{\link{plot.LikeGridSearch}()}
}
\concept{FCVAR auxilliary functions}
