% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{LMtest}
\alias{LMtest}
\title{Breusch-Godfrey Lagrange Multiplier Test for Serial Correlation}
\usage{
LMtest(x, q)
}
\arguments{
\item{x}{A vector or Tx1 matrix of variables to be tested,
typically model residuals.}

\item{q}{The number of lags for the serial correlation tests.}
}
\value{
A list object \code{LMtest_out} containing the test results,
including the following parameters:
\describe{
  \item{\code{LM}}{The LM statistic for individual series.}
  \item{\code{pv}}{The p-value for LM-test on individual series.}
}
}
\description{
\code{LMtest} performs a Breusch-Godfrey Lagrange Multiplier test
for serial correlation.
}
\note{
The LM test is consistent for heteroskedastic series.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
MVWNtest_stats <- MVWNtest(x = results$Residuals, maxlag = 12, printResults = 1)
LMtest(x = matrix(results$Residuals[, 1]), q = 12)
LMtest(x = results$Residuals[,2, drop = FALSE], q = 12)

set.seed(27)
WN <- stats::rnorm(100)
RW <- cumsum(stats::rnorm(100))
LMtest(x = matrix(WN), q = 10)
LMtest(x = matrix(RW), q = 10)
MVWN_x <- as.matrix(data.frame(WN = WN, RW = RW))
MVWNtest_stats <- MVWNtest(x = MVWN_x, maxlag = 10, printResults = 1)
}
\seealso{
\code{MVWNtest} calls this function to test residuals
from the estimation results of \code{FCVARestn}.
An alternative test is the Ljung-Box Q-test in \code{Qtest}.

Other FCVAR postestimation functions: 
\code{\link{FCVARboot}()},
\code{\link{FCVARhypoTest}()},
\code{\link{GetCharPolyRoots}()},
\code{\link{MVWNtest}()},
\code{\link{Qtest}()},
\code{\link{plot.GetCharPolyRoots}()},
\code{\link{print.GetCharPolyRoots}()},
\code{\link{print.MVWNtest}()}
}
\concept{FCVAR postestimation functions}
