% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{mv_wntest}
\alias{mv_wntest}
\title{Multivariate White Noise Tests}
\usage{
mv_wntest(x, maxlag, printResults)
}
\arguments{
\item{x}{A matrix of variables to be included in the system,
typically model residuals.}

\item{maxlag}{The number of lags for serial correlation tests.}

\item{printResults}{An indicator to print results to screen.}
}
\value{
A list object \code{mv_wntest_out} containing the test results,
including the following parameters:
\describe{
  \item{\code{Q}}{A 1xp vector of Q statistics for individual series.}
  \item{\code{pvQ}}{A 1xp vector of P-values for Q-test on individual series.}
  \item{\code{LM}}{A 1xp vector of LM statistics for individual series.}
  \item{\code{pvLM}}{A 1xp vector of P-values for LM-test on individual series.}
  \item{\code{mvQ}}{A multivariate Q statistic.}
  \item{\code{pvMVQ}}{A p-value for multivariate Q-statistic using \code{p^2*maxlag}
  degrees of freedom.}
}
}
\description{
\code{mv_wntest} performs multivariate tests for white noise.
It performs both the Ljung-Box Q-test and the LM-test on individual series
for a sequence of lag lengths.
}
\note{
The LM test should be consistent for heteroskedastic series, Q-test is not.
}
\examples{
opt <- EstOptions()
x <- data(JNP2014)
results <- FCVARestn(x,k = 3,r = 1, opt)
mv_wntest(x = results$Residuals, maxlag = 12, printResults = 1)

mv_wntest(x = rnorm(100), maxlag = 10, printResults = 1)

mv_wntest(x = cumsum(rnorm(100)), maxlag = 10, printResults = 1)
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} produces the residuals intended for this test.
\code{LagSelect} uses this test as part of the lag order selection process.

Other FCVAR postestimation functions: 
\code{\link{HypoTest}()},
\code{\link{LMtest}()},
\code{\link{Qtest}()}
}
\concept{FCVAR postestimation functions}
