% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{MVWNtest}
\alias{MVWNtest}
\title{Multivariate White Noise Tests}
\usage{
MVWNtest(x, maxlag, printResults)
}
\arguments{
\item{x}{A matrix of variables to be included in the system,
typically model residuals.}

\item{maxlag}{The number of lags for serial correlation tests.}

\item{printResults}{An indicator to print results to screen.}
}
\value{
A list object \code{MVWNtest_out} containing the test results,
including the following parameters:
\describe{
  \item{\code{Q}}{A 1xp vector of Q statistics for individual series.}
  \item{\code{pvQ}}{A 1xp vector of P-values for Q-test on individual series.}
  \item{\code{LM}}{A 1xp vector of LM statistics for individual series.}
  \item{\code{pvLM}}{A 1xp vector of P-values for LM-test on individual series.}
  \item{\code{mvQ}}{A multivariate Q statistic.}
  \item{\code{pvMVQ}}{A p-value for multivariate Q-statistic using \code{p^2*maxlag}
  degrees of freedom.}
}
}
\description{
\code{MVWNtest} performs multivariate tests for white noise.
It performs both the Ljung-Box Q-test and the LM-test on individual series
for a sequence of lag lengths.
}
\note{
The LM test should be consistent for heteroskedastic series, Q-test is not.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
MVWNtest_out <- MVWNtest(x = results$Residuals, maxlag = 12, printResults = 1)

set.seed(27)
WN <- rnorm(100)
RW <- cumsum(rnorm(100))
MVWN_x <- as.matrix(data.frame(WN = WN, RW = RW))
MVWNtest_out <- MVWNtest(x = MVWN_x, maxlag = 10, printResults = 1)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} produces the residuals intended for this test.
\code{LagSelect} uses this test as part of the lag order selection process.

Other FCVAR postestimation functions: 
\code{\link{FCVARboot}()},
\code{\link{GetCharPolyRoots}()},
\code{\link{HypoTest}()},
\code{\link{LMtest}()},
\code{\link{Qtest}()}
}
\concept{FCVAR postestimation functions}
