% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{plot.FCVARlikeGrid}
\alias{plot.FCVARlikeGrid}
\title{Plot the Likelihood Function for the FCVAR Model}
\usage{
\method{plot}{FCVARlikeGrid}(likeGrid_params, k, r, opt, file = NULL, file_ext = NULL, main = NULL)
}
\arguments{
\item{likeGrid_params}{A list output from \code{FCVARlikeGrid}.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}

\item{file}{A string path and file name in which to plot a figure.
Renders to the plot window if running in RStudio if NULL.}

\item{file_ext}{A string file extension to indicate the graphics format.
Either png or pdf are currently supported.}

\item{main}{The main title of the plot, passed to \code{plot}.
If \code{main == 'default'}, a generic title is used.}
}
\description{
\code{plot.FCVARlikeGrid} plots the likelihood function from \code{FCVARlikeGrid}.
\code{FCVARlikeGrid} performs a grid-search optimization
by calculating the likelihood function
on a grid of candidate parameter values.
This function evaluates the likelihood over a grid of values
	for \code{c(d,b)} (or \code{phi}).
	It can be used when parameter estimates are sensitive to
	starting values to give an approximation of the global max which can
	then be used as the starting value in the numerical optimization in
	\code{FCVARestn}.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
opt$progress <- 2 # Show progress report on each value of b.
newOpt <- FCVARoptionUpdates(opt, p = 3, r = 1) # Need to update restriction matrices.
likeGrid_params <- FCVARlikeGrid(x, k = 2, r = 1, newOpt)
plot.FCVARlikeGrid(likeGrid_params, k = 2, r = 1, newOpt, main = 'default')
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{plot.FCVARlikeGrid} plots the likelihood function from \code{FCVARlikeGrid}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeGrid}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetEstimates}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{Lbk}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
