% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{plot.GetCharPolyRoots}
\alias{plot.GetCharPolyRoots}
\title{Plot Roots of the Characteristic Polynomial}
\usage{
\method{plot}{GetCharPolyRoots}(
  cPolyRoots,
  b,
  file = NULL,
  file_ext = NULL,
  xlim = NULL,
  ylim = NULL,
  main = NULL
)
}
\arguments{
\item{cPolyRoots}{A vector of the roots of the characteristic polynomial.
An element of the list of estimation \code{results} output from \code{FCVARestn}.}

\item{b}{The order of fractional integration.}

\item{file}{A string path and file name in which to plot a figure.
Renders to the plot window if running in RStudio if NULL.}

\item{file_ext}{A string file extension to indicate the graphich format.
Either png or pdf are currently supported.}

\item{xlim}{The coordinates for the horizontal limits of the axes, passed to \code{plot},
otherwise set to double the maximum magnitude of points on the unit circle
or the transformed unit circle.}

\item{ylim}{The coordinates for the vertical limits of the axes, passed to \code{plot},
otherwise set to double the maximum magnitude of points on the unit circle
or the transformed unit circle.}

\item{main}{The main title of the plot, passed to \code{plot}.
If \code{main == 'default'}, a generic title is used.}
}
\description{
\code{plot.GetCharPolyRoots} plots the output of
\code{GetCharPolyRoots} to screen or to a file.
\code{GetCharPolyRoots} calculates the roots of the
characteristic polynomial and plots them with the unit circle
transformed for the fractional model, see Johansen (2008).
}
\note{
The roots are calculated from the companion form of the VAR,
where the roots are given as the inverse eigenvalues of the
coefficient matrix.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
cPolyRoots <- GetCharPolyRoots(results$coeffs, opt, k = 2, r = 1, p = 3)
print.GetCharPolyRoots(cPolyRoots)
plot.GetCharPolyRoots(cPolyRoots, b = results$coeffs$db[2])
}
\references{
Johansen, S. (2008). "A representation theory for a class of
vector autoregressive models for fractional processes,"
Econometric Theory 24, 651-676.
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} to estimate the model for which to calculate the roots
of the characteristic polynomial.
\code{print.GetCharPolyRoots} prints the output of
\code{GetCharPolyRoots} to screen.

Other FCVAR postestimation functions: 
\code{\link{FCVARboot}()},
\code{\link{GetCharPolyRoots}()},
\code{\link{HypoTest}()},
\code{\link{LMtest}()},
\code{\link{MVWNtest}()},
\code{\link{Qtest}()},
\code{\link{print.GetCharPolyRoots}()},
\code{\link{print.MVWNtest}()}
}
\concept{FCVAR postestimation functions}
