% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_estn.R
\name{print.FCVARestn}
\alias{print.FCVARestn}
\title{Print Estimation Results from the FCVAR model}
\usage{
\method{print}{FCVARestn}(results, k, r, p, T, opt)
}
\arguments{
\item{results}{A list object containing the estimation results of \code{FCVARestn}.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{p}{The number of variables in the system.}

\item{T}{The sample size.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\description{
\code{print.FCVARestn} prints the estimation results from
the output of \code{FCVARestn}.
\code{FCVARestn} estimates the Fractionally Cointegrated VAR model.
 It is the central function in the \code{FCVAR} package with several nested functions, each
	described below. It estimates the model parameters, calculates the
	standard errors and the number of free parameters, obtains the residuals
	and the roots of the characteristic polynomial.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
FCVARestn(x, k = 2, r = 1, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} calls this function at the start of each estimation to verify
validity of options.
\code{print.FCVARestn} prints the output of \code{FCVARestn} to screen.

Other FCVAR estimation functions: 
\code{\link{FCVARestn}()},
\code{\link{FCVARoptionUpdates}()},
\code{\link{FCVARoptions}()},
\code{\link{GetBounds}()}
}
\concept{FCVAR estimation functions}
