% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_spec.R
\name{print.FCVARrankTests}
\alias{print.FCVARrankTests}
\title{Print Statistics from Tests for Cointegrating Rank}
\usage{
\method{print}{FCVARrankTests}(stats, k, p, T, opt)
}
\arguments{
\item{stats}{A list object \code{rankTestStats} containing the results
from repeated estimation of the FCVAR model with different
cointegrating ranks. It is the output of \code{FCVARrankTests}.}

\item{k}{The number of lags in the system.}

\item{p}{The number of variables in the system.}

\item{T}{The sample size.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\description{
\code{print.FCVARrankTests} prints the table of statistics from
the output of \code{FCVARrankTests}.
\code{FCVARrankTests} performs a sequence of  likelihood ratio tests
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
rankTestStats <- FCVARrankTests(x, k = 2, opt)
print.FCVARrankTests(stats = rankTestStats, k = 2, p = ncol(x), T = nrow(x), opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} is called repeatedly within this function
for each candidate cointegrating rank.
\code{print.FCVARrankTests} prints the output of \code{FCVARrankTests} to screen.

Other FCVAR specification functions: 
\code{\link{FCVARbootRank}()},
\code{\link{FCVARlagSelect}()},
\code{\link{FCVARrankTests}()},
\code{\link{print.FCVARlagSelect}()}
}
\concept{FCVAR specification functions}
