% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_spec.R
\name{print.LagSelect}
\alias{print.LagSelect}
\title{Print Statistics from Lag Order Selection}
\usage{
\method{print}{LagSelect}(stats, kmax, r, p, T, order, opt)
}
\arguments{
\item{stats}{A list object \code{LagSelectStats} containing the results
from repeated estimation of the FCVAR model with different orders
of the autoregressive lag length. It is the output of \code{LagSelect}.}

\item{kmax}{The maximum number of lags in the system.}

\item{r}{The cointegrating rank. This is often set equal to \code{p},
the number of variables in the system, since it is better to overspecify
than underspecify the model.}

\item{p}{The number of variables in the system.}

\item{T}{The sample size.}

\item{order}{The order of serial correlation for white noise tests.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\description{
\code{print.LagSelect} prints the table of statistics from
the output of \code{LagSelect}.
\code{LagSelect} takes a matrix of variables and performs lag
	selection on it by using the likelihood ratio test.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
LagSelectStats <- LagSelect(x, kmax = 3, r = 3, order = 12, opt)
print.LagSelect(stats = LagSelectStats, kmax = 3, r = 3, p = 3, order, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} is called repeatedly within this function
for each candidate lag order.
\code{print.LagSelect} prints the output of \code{LagSelect} to screen.

Other FCVAR specification functions: 
\code{\link{FCVARbootRank}()},
\code{\link{FCVARrankTests}()},
\code{\link{LagSelect}()},
\code{\link{print.FCVARrankTests}()}
}
\concept{FCVAR specification functions}
