% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{print.MVWNtest}
\alias{print.MVWNtest}
\title{Print Statistics for Multivariate White Noise Tests}
\usage{
\method{print}{MVWNtest}(stats, maxlag, p)
}
\arguments{
\item{stats}{A list object \code{MVWNtest_stats} containing the results
from multivariate tests for white noise
It is the output of \code{MVWNtest}.}

\item{maxlag}{The number of lags for serial correlation tests.}

\item{p}{The number of variables in the system.}
}
\description{
\code{print.MVWNtest} prints the statistics from \code{MVWNtest} to screen.
\code{MVWNtest} performs multivariate tests for white noise.
It performs both the Ljung-Box Q-test and the LM-test on individual series
for a sequence of lag lengths.
}
\note{
The LM test should be consistent for heteroskedastic series, Q-test is not.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
MVWNtest_stats <- MVWNtest(x = results$Residuals, maxlag = 12, printResults = 1)
\dontrun{print.MVWNtest(stats = MVWNtest_stats, maxlag = 12, p = 3)}

set.seed(27)
WN <- stats::rnorm(100)
RW <- cumsum(stats::rnorm(100))
MVWN_x <- as.matrix(data.frame(WN = WN, RW = RW))
MVWNtest_stats <- MVWNtest(x = MVWN_x, maxlag = 10, printResults = 1)
\dontrun{print.MVWNtest(stats = MVWNtest_stats, maxlag = 10, p = 2)}
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} produces the residuals intended for this test.
\code{LagSelect} uses this test as part of the lag order selection process.
\code{print.MVWNtest} prints the output of \code{MVWNtest} to screen.

Other FCVAR postestimation functions: 
\code{\link{FCVARboot}()},
\code{\link{FCVARhypoTest}()},
\code{\link{GetCharPolyRoots}()},
\code{\link{LMtest}()},
\code{\link{MVWNtest}()},
\code{\link{Qtest}()},
\code{\link{plot.GetCharPolyRoots}()},
\code{\link{print.GetCharPolyRoots}()}
}
\concept{FCVAR postestimation functions}
