% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{Qtest}
\alias{Qtest}
\title{Ljung-Box Q-test for Serial Correlation}
\usage{
Qtest(x, maxlag)
}
\arguments{
\item{x}{A vector or Tx1 matrix of variables to be tested,
typically model residuals.}

\item{maxlag}{The number of lags for the serial correlation tests.}
}
\value{
A list object \code{LMtest_out} containing the test results,
including the following parameters:
\describe{
  \item{\code{Qstat}}{A 1xp vector of Q statistics for individual series.}
  \item{\code{pv}}{A 1xp vector of P-values for Q-test on individual series.}
}
}
\description{
\code{Qtest} performs a (multivariate) Ljung-Box Q-test for serial correlation, see
}
\note{
The LM test in \code{LMtest} is consistent for heteroskedastic series,
while the Q-test is not.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
MVWNtest_stats <- MVWNtest(x = results$Residuals, maxlag = 12, printResults = 1)
Qtest(x = results$Residuals, maxlag = 12)
Qtest(x = matrix(results$Residuals[, 1]), maxlag = 12)
Qtest(x = results$Residuals[,2, drop = FALSE], maxlag = 12)

set.seed(27)
WN <- rnorm(100)
RW <- cumsum(rnorm(100))
MVWN_x <- as.matrix(data.frame(WN = WN, RW = RW))
Qtest(x = MVWN_x, maxlag = 10)
Qtest(x = matrix(WN), maxlag = 10)
Qtest(x = matrix(RW), maxlag = 10)
}
\references{
H. Luetkepohl (2005) "New Introduction to Multiple Time Series Analysis," Springer, Berlin.
}
\seealso{
\code{MVWNtest} calls this function to test residuals
from the estimation results of \code{FCVARestn}.
An alternative test is the Breusch-Godfrey Lagrange Multiplier Test in \code{LMtest}.

Other FCVAR postestimation functions: 
\code{\link{FCVARboot}()},
\code{\link{GetCharPolyRoots}()},
\code{\link{HypoTest}()},
\code{\link{LMtest}()},
\code{\link{MVWNtest}()},
\code{\link{plot.GetCharPolyRoots}()},
\code{\link{print.GetCharPolyRoots}()},
\code{\link{print.MVWNtest}()}
}
\concept{FCVAR postestimation functions}
