% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_spec.R
\name{RankTests}
\alias{RankTests}
\title{Test for Cointegrating Rank}
\usage{
RankTests(x, k, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{EstOptions()}.}
}
\value{
A list object \code{rankTestStats} containing the results
from cointegrating rank tests, containing the following \code{(p+1)} vectors
with \code{i}th element corresponding to \code{rank = i-1}:
including the following parameters:
\describe{
  \item{\code{dHat}}{Estimates of \code{d}.}
  \item{\code{bHat}}{Estimates of \code{b}.}
  \item{\code{LogL}}{Maximized log-likelihood.}
  \item{\code{LRstat}}{LR trace statistic for testing rank \code{r} against rank \code{p}.}
  \item{\code{pv}}{The p-value of LR trace test, or "999" if p-value is not available.}
}
}
\description{
\code{RankTests} performs a sequence of  likelihood ratio tests
}
\examples{
opt <- EstOptions()
x <- data(JNP2014)
RankTests(x, k = 2, opt)
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} is called repeatedly within this function
for each candidate cointegrating rank.

Other FCVAR specification functions: 
\code{\link{LagSelect}()},
\code{\link{get_pvalues}()}
}
\concept{FCVAR specification functions}
