% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{SEmat2vecU}
\alias{SEmat2vecU}
\title{Collect Parameters into a Vector}
\usage{
SEmat2vecU(coeffs, k, r, p, opt)
}
\arguments{
\item{coeffs}{A list of coefficients of the FCVAR model.
It is an element of the list \code{results} returned by \code{FCVARestn}.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{p}{The number of variables in the system.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{EstOptions()}.}
}
\value{
A vector \code{param} of parameters in the FCVAR model.
}
\description{
\code{SEmat2vecU} transforms the model parameters in matrix
	form into a vector.
}
\examples{
opt <- EstOptions()
x <- data(JNP2014)
results <- FCVARestn(x, k = 2, r = 1, opt)
param <- SEmat2vecU(coeffs = results$coeffs, k = 2, r = 1, p = 3, opt)
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} to estimate the FCVAR model and calculate
standard errors of the estimates.
\code{SEmat2vecU} is called by \code{FCVARhess} to sort the parameters
into a vector to calculate the Hessian matrix.
\code{SEvec2matU} is the inverse of \code{SEmat2vecU}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{FreeParams}()},
\code{\link{FullFCVARlike}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{RstrctOptm_Switch}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
