% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{SEvec2matU}
\alias{SEvec2matU}
\title{Extract Parameters from a Vector}
\usage{
SEvec2matU(param, k, r, p, opt)
}
\arguments{
\item{param}{A vector of parameters in the FCVAR model.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{p}{The number of variables in the system.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
\code{coeffs}, a list of coefficients of the FCVAR model.
It has the same form as an element of the list \code{results}
returned by \code{FCVARestn}.
}
\description{
\code{SEvec2matU} transforms the vectorized model parameters
	into matrices.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
params <- SEmat2vecU(coeffs = results$coeffs, k = 2, r = 1, p = 3, opt)
coeffs <- SEvec2matU(param = params, k = 2, r = 1, p = 3, opt )

params <- matrix(seq(25))
coeffs <- SEvec2matU(param = params, k = 2, r = 1, p = 3, opt )
params <- SEmat2vecU(coeffs = coeffs, k = 2, r = 1, p = 3, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} to estimate the FCVAR model and calculate
standard errors of the estimates.
\code{SEmat2vecU} is called by \code{FCVARhess} to convert the parameters
from a vector into the coefficients after calculating the Hessian matrix.
\code{SEmat2vecU} is a near inverse of \code{SEvec2matU},
in the sense that \code{SEvec2matU} obtains only a
subset of the parameters in \code{results$coeffs}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{Lbk}()},
\code{\link{LikeGridSearch}()},
\code{\link{SEmat2vecU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
