% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{TransformData}
\alias{TransformData}
\title{Transform Data for Regression}
\usage{
TransformData(x, k, db, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{db}{The orders of fractional integration.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A list object \code{Z_array} containing the transformed data,
including the following parameters:
\describe{
  \item{\code{Z0}}{A matrix of data calculated by fractionally differencing
  \code{x} at differencing order \code{d}.}
  \item{\code{Z1}}{The matrix \code{x} (augmented with a vector of ones
  if the model includes a restricted constant term), which is then lagged and stacked
  and fractionally differenced (with order \eqn{d-b}).}
  \item{\code{Z2}}{The matrix \code{x} lagged and stacked
  and fractionally differenced (with order \code{d}).}
  \item{\code{Z3}}{A column of ones if model includes an
  unrestricted constant term, otherwise \code{NULL}.}
}
}
\description{
\code{TransformData} transforms the raw data by fractional differencing.
The output is in the format required for regression and
reduced rank regression.
}
\examples{
opt <- FCVARoptions()
x <- data(JNP2014)
Z_array <- TransformData(x, k = 2, db = c(1, 1), opt)
}
\references{
Johansen, S. and M. \enc{Ø}{O}. Nielsen (2012).
"Likelihood inference for a fractionally cointegrated
vector autoregressive model," Econometrica 80, 2667-2732.

Johansen, S. (1995). "Likelihood-Based Inference
in Cointegrated Vector Autoregressive Models,"
New York: Oxford University Press.
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} calls \code{GetParams}, which calls \code{TransformData}
to estimate the FCVAR model.
\code{TransformData} in turn calls \code{FracDiff} and \code{Lbk}
to perform the transformation.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{FreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{RstrctOptm_Switch}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()}
}
\concept{FCVAR auxilliary functions}
