% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_opt.R
\name{updateRestrictions}
\alias{updateRestrictions}
\title{Update FCVAR Restrictions}
\usage{
updateRestrictions(opt, p, r)
}
\arguments{
\item{opt}{A list object that stores the chosen estimation options,
generated from \code{EstOptions()}.}

\item{p}{The number of variables in the system.}

\item{r}{The cointegrating rank.}
}
\value{
A list object \code{newOpt} that stores the chosen estimation options,
with appropriate updates based on user-defined options.
}
\description{
A function to set and test estimation options for validity and compatibility.
This function is called prior to estimation and
	performs several checks to verify the input in the estimation
	options:
\itemize{
  \item Check that only one option for deterministics is chosen,
  \item Check that starting values have been specified correctly,
  \item Update \code{dbMin}, \code{dbMax}, based on user-specified options,
  \item Check for appropriate dimensions and redundancies in the
}
}
\examples{
opt <- EstOptions()
opt$gridSearch <- 0 # Skip grid search in optimization.
newOpt <- updateRestrictions(opt, p = 3, r = 1)
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} calls this function at the start of each estimation to verify
validity of options.

Other FCVAR option functions: 
\code{\link{EstOptions}()}
}
\concept{FCVAR option functions}
