% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR.R
\docType{data}
\name{votingJNP2014}
\alias{votingJNP2014}
\title{Aggregate support for Canadian political parties.}
\format{
A data frame with 316 rows and 6 variables:
\describe{
  \item{lib}{aggregate support for the Liberal party}
  \item{pc}{aggregate support for the Conservative party}
  \item{ir_can}{Canadian 3-month T-bill rates}
  \item{ir_us}{US 3-month T-bill rates}
  \item{un_can}{Canadian unemployment rate}
  \item{un_us}{US unemployment rate}
}
}
\source{
\url{https://sites.google.com/view/mortennielsen/software}
}
\usage{
votingJNP2014
}
\description{
A dataset containing the aggregate support for Canadian political
parties and economic indicators from Canada and the United States.
}
\keyword{datasets}
