% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR.R
\docType{package}
\name{FCVAR}
\alias{FCVAR}
\title{FCVAR: A package for estimating the Fractionally Cointegrated VAR model.}
\description{
The FCVAR package provides three categories of important functions:
Estimation, Postestimation, Specificationand Auxilliary functions.
}
\section{Estimation functions}{

The estimation functions include the main estimation function \code{FCVARestn}
  and associated functions to display estimation results.
  These include:
\describe{
  \item{\code{First_function}}{This is the first function}
  \item{\code{Second_function}}{This is the second function}
}
}

\section{Postestimation functions}{

The postestimation functions are used to display the results and test the
  the goodness of fit of the estimated model.
  These include:
\describe{
  \item{\code{First_function}}{This is the first function}
  \item{\code{Second_function}}{This is the second function}
}
}

\section{Specification functions}{

The specification functions are used to estimate a series of models in order
  to make model specfication decisions.
  These include:
\describe{
  \item{\code{First_function}}{This is the first function}
  \item{\code{Second_function}}{This is the second function}
}
}

\section{Auxilliary functions}{

The auxilliary functions are used to perform intermediate calculations
  for the estimation function.
  Most of these functions are designed for use only within the estimation function.
  Selected functions for external usage include:
\describe{
  \item{\code{First_function}}{This is the first function}
  \item{\code{Second_function}}{This is the second function}
}
}

\section{Examples}{

Several datasets are included as examples of the model building process.
  These include \code{JNP2014} and \code{BDP1997}.
  Sample model builds with hypothesis tests are found in the example scripts
  \code{JNPexample} and \code{BDPexample}.
}

