% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{FCVARhess}
\alias{FCVARhess}
\title{Calculate the Hessian Matrix}
\usage{
FCVARhess(x, k, r, coeffs, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{coeffs}{A list of coefficients of the FCVAR model.
It is an element of the list \code{results} returned by \code{FCVARestn}.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
The \code{hessian} matrix  of second derivatives of the FCVAR
log-likelihood function, calculated with the parameter estimates
specified in \code{coeffs}.
}
\description{
\code{FCVARhess} calculates the Hessian matrix of the
	log-likelihood by taking numerical derivatives.
	It is used to calculate the standard errors of parameter estimates.
}
\examples{
opt <- FCVARoptions()
x <- data(JNP2014)
results <- FCVARestn(x, k = 2, r = 1, opt)
hessian <- FCVARhess(x, k = 2, r = 1, coeffs = results$coeffs, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} to estimate the FCVAR model and calculate
standard errors of the estimates.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{FreeParams}()},
\code{\link{FullFCVARlike}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{RstrctOptm_Switch}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
