% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{GetCharPolyRoots}
\alias{GetCharPolyRoots}
\title{Roots of the Characteristic Polynomial}
\usage{
GetCharPolyRoots(coeffs, opt, k, r, p)
}
\arguments{
\item{coeffs}{A list of coefficients for the FCVAR model.
An element of the list of estimation \code{results} output from \code{FCVARestn}.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{p}{The number of variables in the system.}
}
\value{
A complex vector \code{cPolyRoots} with the roots of the characteristic polynomial.
}
\description{
\code{GetCharPolyRoots} calculates the roots of the
characteristic polynomial and plots them with the unit circle
transformed for the fractional model, see Johansen (2008).
}
\note{
The roots are calculated from the companion form of the VAR,
where the roots are given as the inverse eigenvalues of the
coefficient matrix.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
cPolyRoots <- GetCharPolyRoots(results$coeffs, opt, k = 2, r = 1, p = 3)
}
\references{
Johansen, S. (2008). "A representation theory for a class of
vector autoregressive models for fractional processes,"
Econometric Theory 24, 651-676.
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} to estimate the model for which to calculate the roots
of the characteristic polynomial.

Other FCVAR postestimation functions: 
\code{\link{FCVARboot}()},
\code{\link{HypoTest}()},
\code{\link{LMtest}()},
\code{\link{MVWNtest}()},
\code{\link{Qtest}()},
\code{\link{print.MVWNtest}()}
}
\concept{FCVAR postestimation functions}
