% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{GetResiduals}
\alias{GetResiduals}
\title{Calculate Residuals for the FCVAR Model}
\usage{
GetResiduals(x, k, r, coeffs, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{coeffs}{A list of coefficients of the FCVAR model.
It is an element of the list \code{results} returned by \code{FCVARestn}.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{EstOptions()}.}
}
\value{
A matrix \code{epsilon} of residuals from FCVAR model estimation
calculated with the parameter estimates specified in \code{coeffs}.
}
\description{
\code{GetResiduals} calculate residuals for the FCVAR model
from given parameter values.
}
\examples{
opt <- EstOptions()
x <- data(JNP2014)
results <- FCVARestn(x, k = 2, r = 1, opt)
epsilon <- GetResiduals(x, k = 2, r = 1, coeffs = results$coeffs, opt)
}
\references{
Johansen, S. and M. \enc{Ø}{O}. Nielsen (2012).
"Likelihood inference for a fractionally cointegrated
vector autoregressive model," Econometrica 80, 2667-2732.
}
\seealso{
\code{EstOptions} to set default estimation options.
\code{FCVARestn} to estimate the FCVAR model.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FracDiff}()},
\code{\link{FreeParams}()},
\code{\link{FullFCVARlike}()},
\code{\link{GetParams}()},
\code{\link{Lbk}()},
\code{\link{LikeGrid}()},
\code{\link{TransformData}()}
}
\concept{FCVAR auxilliary functions}
