% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_spec.R
\name{LagSelect}
\alias{LagSelect}
\title{Select Lag Order}
\usage{
LagSelect(x, kmax, r, order, opt)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{kmax}{The maximum number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{order}{The order of serial correlation for white noise tests.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\description{
\code{LagSelect} takes a matrix of variables and performs lag
	selection on it by using the likelihood ratio test. Output and test
	results are printed to the screen.
}
\examples{
opt <- FCVARoptions()
x <- data(JNP2014)
LagSelect(x, kmax = 4, r = 3, order = 12, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} is called repeatedly within this function
for each candidate lag order.

Other FCVAR specification functions: 
\code{\link{FCVARbootRank}()},
\code{\link{GetPvalues}()},
\code{\link{RankTests}()}
}
\concept{FCVAR specification functions}
