% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{Lbk}
\alias{Lbk}
\title{Calculate Lag Polynomial in the Fractional Lag Operator}
\usage{
Lbk(x, b, k)
}
\arguments{
\item{x}{A matrix of variables to be included in the system.}

\item{b}{The order of fractional differencing.}

\item{k}{The number of lags in the system.}
}
\value{
A matrix \code{Lbkx} of the form \eqn{[ Lb^1 x, Lb^2 x, ..., Lb^k x]}
where \eqn{Lb = 1 - (1-L)^b}.
The output matrix has the same number of rows as \code{x}
but \code{k} times as many columns.
}
\description{
\code{Lbk} calculates a lag polynomial in the fractional lag operator.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
Lbkx <- Lbk(x, b = results$coeffs$db[2], k = 2)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} calls \code{GetParams}, which calls \code{TransformData}
to estimate the FCVAR model.
\code{TransformData} in turn calls \code{FracDiff} and \code{Lbk}
to perform the transformation.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeGrid}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetEstimates}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{SEmat2vecU}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()},
\code{\link{plot.FCVAR_grid}()}
}
\concept{FCVAR auxilliary functions}
