% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{LMtest}
\alias{LMtest}
\title{Breusch-Godfrey Lagrange Multiplier Test for Serial Correlation}
\usage{
LMtest(x, q)
}
\arguments{
\item{x}{A vector or Tx1 matrix of variables to be tested,
typically model residuals.}

\item{q}{The number of lags for the serial correlation tests.}
}
\value{
A list object \code{LMtest_out} containing the test results,
including the following parameters:
\describe{
  \item{\code{LM}}{The LM statistic for individual series.}
  \item{\code{pv}}{The p-value for LM-test on individual series.}
}
}
\description{
\code{LMtest} performs a Breusch-Godfrey Lagrange Multiplier test
for serial correlation.
}
\note{
The LM test is consistent for heteroskedastic series.
}
\examples{
LMtest(x = rnorm(100), q = 10)

LMtest(x = cumsum(rnorm(100)), q = 10)
}
\seealso{
\code{mv_wntest} calls this function to test residuals
from the estimation results of \code{FCVARestn}.
An alternative test is the Ljung-Box Q-test in \code{Qtest}.

Other FCVAR postestimation functions: 
\code{\link{Qtest}()},
\code{\link{mv_wntest}()}
}
\concept{FCVAR postestimation functions}
