% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_post.R
\name{Qtest}
\alias{Qtest}
\title{Ljung-Box Q-test for Serial Correlation}
\usage{
Qtest(x, maxlag)
}
\arguments{
\item{x}{A vector or Tx1 matrix of variables to be tested,
typically model residuals.}

\item{maxlag}{The number of lags for the serial correlation tests.}
}
\value{
A list object \code{LMtest_out} containing the test results,
including the following parameters:
\describe{
  \item{\code{Qstat}}{A 1xp vector of Q statistics for individual series.}
  \item{\code{pv}}{A 1xp vector of P-values for Q-test on individual series.}
}
}
\description{
\code{Qtest} performs a (Multivariate) Ljung-Box Q-test for serial correlation, see
}
\note{
The LM test in \code{LMtest} is consistent for heteroskedastic series,
while the Q-test is not.
}
\examples{
Qtest(x = rnorm(100), maxlag = 10)

Qtest(x = cumsum(rnorm(100)), maxlag = 10)
}
\references{
H. Luetkepohl (2005) "New Introduction to Multiple Time Series Analysis," Springer, Berlin.
}
\seealso{
\code{mv_wntest} calls this function to test residuals
from the estimation results of \code{FCVARestn}.
An alternative test is the Breusch-Godfrey Lagrange Multiplier Test in \code{LMtest}.

Other FCVAR postestimation functions: 
\code{\link{LMtest}()},
\code{\link{mv_wntest}()}
}
\concept{FCVAR postestimation functions}
