% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCVAR_aux.R
\name{SEmat2vecU}
\alias{SEmat2vecU}
\title{Collect Parameters into a Vector}
\usage{
SEmat2vecU(coeffs, k, r, p, opt)
}
\arguments{
\item{coeffs}{A list of coefficients of the FCVAR model.
It is an element of the list \code{results} returned by \code{FCVARestn}.}

\item{k}{The number of lags in the system.}

\item{r}{The cointegrating rank.}

\item{p}{The number of variables in the system.}

\item{opt}{A list object that stores the chosen estimation options,
generated from \code{FCVARoptions()}.}
}
\value{
A vector \code{param} of parameters in the FCVAR model.
}
\description{
\code{SEmat2vecU} transforms the model parameters in matrix
	form into a vector.
}
\examples{
opt <- FCVARoptions()
opt$gridSearch   <- 0 # Disable grid search in optimization.
opt$dbMin        <- c(0.01, 0.01) # Set lower bound for d,b.
opt$dbMax        <- c(2.00, 2.00) # Set upper bound for d,b.
opt$constrained  <- 0 # Impose restriction dbMax >= d >= b >= dbMin ? 1 <- yes, 0 <- no.
x <- votingJNP2014[, c("lib", "ir_can", "un_can")]
results <- FCVARestn(x, k = 2, r = 1, opt)
params <- SEmat2vecU(coeffs = results$coeffs, k = 2, r = 1, p = 3, opt)
coeffs <- SEvec2matU(param = params, k = 2, r = 1, p = 3, opt )

params <- matrix(seq(25))
coeffs <- SEvec2matU(param = params, k = 2, r = 1, p = 3, opt )
params <- SEmat2vecU(coeffs = coeffs, k = 2, r = 1, p = 3, opt)
}
\seealso{
\code{FCVARoptions} to set default estimation options.
\code{FCVARestn} to estimate the FCVAR model and calculate
standard errors of the estimates.
\code{SEmat2vecU} is called by \code{FCVARhess} to sort the parameters
into a vector to calculate the Hessian matrix.
\code{SEvec2matU} is a near inverse of \code{SEmat2vecU},
in the sense that \code{SEvec2matU} obtains only a
subset of the parameters in \code{results$coeffs}.

Other FCVAR auxilliary functions: 
\code{\link{FCVARforecast}()},
\code{\link{FCVARhess}()},
\code{\link{FCVARlikeFull}()},
\code{\link{FCVARlikeGrid}()},
\code{\link{FCVARlikeMu}()},
\code{\link{FCVARlike}()},
\code{\link{FCVARsimBS}()},
\code{\link{FCVARsim}()},
\code{\link{FracDiff}()},
\code{\link{GetEstimates}()},
\code{\link{GetFreeParams}()},
\code{\link{GetParams}()},
\code{\link{GetResiduals}()},
\code{\link{GetRestrictedParams}()},
\code{\link{Lbk}()},
\code{\link{SEvec2matU}()},
\code{\link{TransformData}()},
\code{\link{plot.FCVARlikeGrid}()}
}
\concept{FCVAR auxilliary functions}
